requirements.txt
fastapi
uvicorn
sqlalchemy
mysql-connector-python
openai
python-dotenv
httpx
tenacity
streamlit
requests
pydantic